generate config.h in the build directory
So that every build can have its own version of config.h.

Under the current approach, config.h is built under the source
directory, only when config.h does not exist or is older than
config.h.in.

Therefore, if the source directory is shared among multiple
operating systems (e.g. macOS and linux), the first OS that
touches the directory builds config.h, and the other OS reuses the
existing file.

This leads to a build error if the first OS is macOS and the
second is linux, since birthtime_spec is only available in the
former.